+++ pkgs/multistockfish_chess/ios/multistockfish_chess/Sources/multistockfish_chess/Stockfish/src/tune.cpp
@@ -25,6 +25,7 @@
 #include <sstream>
 #include <string>
 
+#include "sfio.h"
 #include "ucioption.h"
 
 using std::string;
@@ -59,12 +60,12 @@ void Tune::make_option(OptionsMap* opts, const string& n, int v, const SetRange&
     LastOption = &((*opts)[n]);
 
     // Print formatted parameters, ready to be copy-pasted in Fishtest
-    std::cout << n << ","                                  //
-              << v << ","                                  //
-              << r(v).first << ","                         //
-              << r(v).second << ","                        //
-              << (r(v).second - r(v).first) / 20.0 << ","  //
-              << "0.0020" << std::endl;
+    sfio::out() << n << ","                                  //
+                << v << ","                                  //
+                << r(v).first << ","                         //
+                << r(v).second << ","                        //
+                << (r(v).second - r(v).first) / 20.0 << ","  //
+                << "0.0020" << std::endl;
 }
 
 string Tune::next(string& names, bool pop) {