+++ pkgs/multistockfish_sf16/ios/multistockfish_sf16/Sources/multistockfish_sf16/Stockfish16/src/misc.cpp
@@ -75,9 +75,9 @@ namespace {
 /// Version number or dev.
 constexpr string_view version = "16";
 
-/// Our fancy logging facility. The trick here is to replace cin.rdbuf() and
-/// cout.rdbuf() with two Tie objects that tie cin and cout to a file stream. We
-/// can toggle the logging of std::cout and std:cin at runtime whilst preserving
+/// Our fancy logging facility. The trick here is to replace the rdbuf() of the
+/// engine's own input and output streams with two Tie objects that tie them to a
+/// file stream. We can toggle the logging at runtime whilst preserving
 /// usual I/O functionality, all without changing a single line of code!
 /// Idea from http://groups.google.com/group/comp.lang.c++/msg/1d941c0f26ea0d81
 
@@ -105,7 +105,7 @@ struct Tie: public streambuf { // MSVC requires split streambuf for cin and cout
 
 class Logger {
 
-  Logger() : in(cin.rdbuf(), file.rdbuf()), out(cout.rdbuf(), file.rdbuf()) {}
+  Logger() : in(sfio::in().rdbuf(), file.rdbuf()), out(sfio::out().rdbuf(), file.rdbuf()) {}
  ~Logger() { start(""); }
 
   ofstream file;
@@ -118,8 +118,8 @@ public:
 
     if (l.file.is_open())
     {
-        cout.rdbuf(l.out.buf);
-        cin.rdbuf(l.in.buf);
+        sfio::out().rdbuf(l.out.buf);
+        sfio::in().rdbuf(l.in.buf);
         l.file.close();
     }
 
@@ -133,8 +133,8 @@ public:
             exit(EXIT_FAILURE);
         }
 
-        cin.rdbuf(&l.in);
-        cout.rdbuf(&l.out);
+        sfio::in().rdbuf(&l.in);
+        sfio::out().rdbuf(&l.out);
     }
   }
 };